backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
statsmodels==0.14.0
yfinance==0.2.31
pandas==2.1.3
python-dotenv==1.0.0
plotly==5.18.0
langgraph==0.0.10
langchain-groq==0.0.1
pmdarima==2.0.4
requests==2.31.0
python-multipart==0.0.6
